Powerups
Added powerups to the game.
They drop 10% of the time when an enemy ship is destroyed.

One of the powerups increases the shield level of the player ship. The maximum shield level is 4.

The second powerup increases the player ship's movement speed by 3. The player ship's movement speed starts at 30.